move truncation out to `solve` so it can be re-used
