+++ src/solve/truncate.rs
